Modifies storeU_id to use Long data type
Changes the storeU_id field in the Users entity and related classes to use the Long data type instead of Integer.

This modification ensures that the application can accommodate a larger range of store IDs, preventing potential overflow issues and providing more flexibility for future scalability.

+++ lab1/src/main/java/com/example/demo/config/Jwt/JwtUtil.java
@@ -20,7 +20,7 @@ public class JwtUtil {
     @Value("${jwt.expiration}")
     private long expiration;
 
-    public String generateToken(String email,  Roles rol, Integer storeU_id) {
+    public String generateToken(String email,  Roles rol, Long storeU_id) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + expiration);
